+++ app/Modules/Scheduler/Controllers/WorkorderController.php
@@ -68,7 +68,10 @@ class WorkorderController extends Controller
 
         event(new WorkorderModified(Workorder::find($workorder->id)));
 
-        //return redirect()->route('workorders.edit', ['id' => $workorder->id])->with('alertSuccess', trans('fi.record_successfully_created'));
-        return back()->with('alertSuccess', trans('fi.record_successfully_created'));
+        if (!$workorder->client->address) {
+            return redirect()->route('workorders.edit', ['id' => $workorder->id]);
+        } else {
+            return back()->with('alertSuccess', trans('fi.record_successfully_created'));
+        }
     }
 }